#6 Determine the date of a slot

+++ backend/src/main/kotlin/the/maltesers/planner/PlannerCount.kt
@@ -3,4 +3,4 @@ package the.maltesers.planner
 import io.micronaut.core.annotation.Introspected
 
 @Introspected
-data class PlannerCount(val total: Int)
+data class PlannerCount(val total: Long)
